+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - added: `autoRotateCouch` helper function.
 - added: `chunkSize` and `startIndex` parameters to rolling database streaming methods.
+- deprecated: `limit` parameter for `viewToStream`. Use `chunkSize` instead.
 
 ## 0.2.23 (2025-09-11)